requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langsmith>=0.2.0
pydantic>=2.0
python-dotenv>=1.0
pytest>=8.0
fastapi>=0.111.0
uvicorn>=0.30.0
